requirements.txt
selenium==4.21.0
pytest==8.3.5
allure-pytest==2.14.2
